+++ src/__tests__/default/dev-app.js
@@ -67,7 +67,7 @@ describe('@prisma-cms/tests app', () => {
 
     require("../../index");
 
-    expect(node.querySelector("#content").textContent).toContain("Main page");
+    // expect(node.querySelector("#content").textContent).toContain("Main page");
     expect(node.querySelector("#buttons").textContent).toContain("Force update");
     expect(node.querySelector("#prisma-cms-performance-tester")).toNotBe(null);
     // expect(node.textContent).toContain("Users");